ai-service/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pandas
requests
python-dotenv
geoip2
pyyaml
gunicorn
